test-requirements.txt
hacking>=0.8.0,<0.9

# Optional backend: SQL
pysqlite

# Optional backend: Memcache
python-memcached>=1.48

# Optional backend: LDAP
# authenticate against an existing LDAP server
python-ldap==2.3.13

# Testing
# computes code coverage percentages
coverage>=3.6
# fixture stubbing
fixtures>=0.3.14
# mock object framework
mock>=1.0
mox>=0.5.3
# required to build documentation
sphinx>=1.1.2,<1.2
# test wsgi apps without starting an http server
WebTest>=2.0

extras
discover
python-subunit
testrepository>=0.0.17
testtools>=0.9.32
testscenarios>=0.4

# for python-keystoneclient
# keystoneclient <0.2.1
httplib2
# replaces httplib2 in keystoneclient >=0.2.1
requests>=1.1
keyring>=1.6.1,<2.0,>=2.1

# For documentation
oslo.sphinx
